refactor: rename folders in src directory and their corresponding namespaces
- We convert their names to lowercase, so:
  1. Editor -> editor
  2. IO -> io
  3. Terminal -> terminal
  4. Utilities -> utilities
  5. Kilo -> kilo
